Move Resolvable into the bag verifier
