+++ bag_verifier/src/main/scala/uk/ac/wellcome/platform/archive/bagverifier/storage/s3/S3Resolvable.scala
@@ -1,9 +1,9 @@
-package uk.ac.wellcome.platform.archive.common.storage.services
+package uk.ac.wellcome.platform.archive.bagverifier.storage.s3
 
 import java.net.URI
 
 import com.amazonaws.services.s3.AmazonS3
-import uk.ac.wellcome.platform.archive.common.storage.Resolvable
+import uk.ac.wellcome.platform.archive.bagverifier.storage.Resolvable
 import uk.ac.wellcome.storage.ObjectLocation
 
 class S3Resolvable(implicit s3Client: AmazonS3)
