+++ bag_verifier/src/test/scala/uk/ac/wellcome/platform/archive/bagverifier/fixity/bag/BagExpectedFixityTest.scala
@@ -5,6 +5,7 @@ import java.net.URI
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
 import uk.ac.wellcome.platform.archive.bagverifier.fixity.ExpectedFileFixity
+import uk.ac.wellcome.platform.archive.bagverifier.storage.Resolvable
 import uk.ac.wellcome.platform.archive.common.bagit.models.{
   BagFetchMetadata,
   BagPath
@@ -13,7 +14,6 @@ import uk.ac.wellcome.platform.archive.common.generators.{
   BagGenerators,
   FetchMetadataGenerators
 }
-import uk.ac.wellcome.platform.archive.common.storage.Resolvable
 import uk.ac.wellcome.platform.archive.common.verify.{
   Checksum,
   ChecksumValue,
